requirements.txt
bottle==0.12.18
gevent==1.4.0
python-ldap==3.2.0
pymongo==3.10.1
